src/requirements.txt
certifi==2020.6.20
chardet==3.0.4
falcon==2.0.0
gunicorn==20.0.4
idna==2.10
psycopg2-binary==2.8.5
PyJWT==1.7.1
pytz==2020.1
requests==2.24.0
six==1.15.0
twilio==6.44.1
urllib3==1.25.10